+++ lib/features/did/did_qr_scan_page.dart
